Cleaned up some of the rules
