+++ buildSrc/src/main/groovy/com/uber/okbuck/bazel/AndroidBinaryRuleComposer.groovy
