+++ buildSrc/src/main/groovy/com/uber/okbuck/bazel/AndroidLibraryRuleComposer.groovy
@@ -24,7 +24,6 @@ final class AndroidLibraryRuleComposer extends AndroidBuckRuleComposer {
                 target.main.sources,
                 target.manifest,
                 providedDeps,
-                "",
                 target.resources)
     }
 }
