+++ buildSrc/src/main/groovy/com/uber/okbuck/bazel/JavaBinaryRuleComposer.groovy
@@ -4,12 +4,11 @@ import com.uber.okbuck.core.model.JavaAppTarget
 import com.uber.okbuck.composer.JavaBuckRuleComposer
 
 final class JavaBinaryRuleComposer extends JavaBuckRuleComposer {
-
-    private JavaBinaryRuleComposer() {
-        // no instance
-    }
-
     static JavaBinaryRule compose(JavaAppTarget target) {
-        return new JavaBinaryRule(bin(target), ["PUBLIC"], [":${src(target)}"], target.mainClass)
+        return new JavaBinaryRule(
+                bin(target),
+                ["//visibility:public"],
+                [":${src(target)}"],
+                target.mainClass)
     }
 }
