+++ buildSrc/src/main/groovy/com/uber/okbuck/bazel/JavaLibraryRuleComposer.groovy
@@ -4,22 +4,16 @@ import com.uber.okbuck.composer.JavaBuckRuleComposer
 import com.uber.okbuck.core.model.JavaLibTarget
 
 final class JavaLibraryRuleComposer extends JavaBuckRuleComposer {
-
-    private JavaLibraryRuleComposer() {
-        // no instance
-    }
-
     static JavaLibraryRule compose(JavaLibTarget target) {
         List<String> deps = []
         deps.addAll(BazelDependencyCache.external(target.main.externalDeps))
         deps.addAll(targets(target.main.targetDeps))
-
         new JavaLibraryRule(
+                "java_library",
                 src(target),
                 ["//visibility:public"],
                 deps,
                 target.main.sources,
                 target.main.resourcesDir)
     }
-
 }
