+++ buildSrc/src/main/resources/META-INF/gradle-plugins/com.uber.okbazel.properties
